Update test utilities.
